refactor(studio.giselles.ai): refactor join functionality with improved error handling and component structure
