--- apps/studio.giselles.ai/app/(auth)/join/utils/redirect-to-error-page.ts
@@ -1,6 +1,0 @@
-import { redirect } from "next/navigation";
-import type { ErrorCode } from "../errors";
-
-export function redirectToErrorPage(token: string, code: ErrorCode) {
-	redirect(`/join/${token}/error/${code}`);
-}
